Refactor services and controllers for user and organization management
- Replaced CompanyService with OrganizationService in service container and related files.
- Updated API endpoints to reflect the change from companies to organizations.
- Created new controllers for user and organization management, including authentication and connection requests.
- Implemented ConnectionService to handle connection requests and connections between organizations.
- Enhanced UserService to manage user signup, login, and profile retrieval.
- Updated frontend components to align with new API endpoints and data structures.
